Move Command-Line aspects to the Terminal project.
