+++ Backend/Backend.csproj
@@ -8,10 +8,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="BetterConsoleTables" Version="2.0.4-rc1" />
-    </ItemGroup>
-
-    <ItemGroup>
       <Compile Remove="Move\Log.cs" />
     </ItemGroup>
