+++ Terminal/Terminal.csproj
@@ -19,6 +19,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <PackageReference Include="BetterConsoleTables" Version="2.0.4-rc1" />
       <PackageReference Include="Hardware.Info" Version="10.0.0.1" />
     </ItemGroup>
